+++ R/data-env.R
@@ -70,7 +70,7 @@
 #' ](https://www.amsterdammedicaldatascience.nl/#amsterdamumcdb)
 #' provided by the Amsterdam University Medical Center, currently is not part
 #' of the core datasets of `ricu`, but code for integrating this dataset is
-#' available on [github](https://github.com/septic-tank/aumc).
+#' available on [github](https://github.com/eth-mds/aumc).
 #'
 #' @section MIMIC-III:
 #' The Medical Information Mart for Intensive Care
@@ -81,8 +81,7 @@
 #' measurements made at the bedside (~1 data point per hour), laboratory test
 #' results, procedures, medications, caregiver notes, imaging reports, and
 #' mortality (both in and out of hospital). For further information, please
-#' refer to the [MIMIC-III documentation
-#' ](https://mimic.physionet.org/about/mimic/).
+#' refer to the [MIMIC-III documentation](https://mimic.physionet.org/iii/).
 #'
 #' The corresponding
 #' [demo dataset](https://physionet.org/content/mimiciii-demo/)