+++ R/data-utils.R
@@ -227,11 +227,15 @@ id_win_helper.hirid_env <- function(x) {
   ids <- field(cfg, "id")
   sta <- field(cfg, "start")
 
-  obs <- as_src_tbl(x, "observations")
-  obs <- subset(obs, .env$ind_fun(.data$patientid, .data$datetime),
-                c("patientid", "datetime"), part_safe = TRUE)
+  obs <- load_src(
+    "observations", x, .env$ind_fun(.data$patientid, .data$datetime),
+    c("patientid", "datetime")
+  )
   obs <- obs[, list(datetime = max(get("datetime"))), by = "patientid"]
-  tbl <- as_src_tbl(x, field(cfg, "table"))[, c(ids, sta)]
+
+  tbl <- load_src(
+    field(cfg, "table"), x, cols = c(ids, sta)
+  )
 
   res <- merge(tbl, obs, by = ids)
   res <- res[, c(sta, "datetime") := lapply(.SD, as_dt_min, get(sta)),
@@ -251,11 +255,12 @@ id_win_helper.aumc_env <- function(x) {
   end <- field(cfg, "end")
 
   tbl <- as_src_tbl(x, unique(field(cfg, "table")))
+
   mis <- setdiff(sta, colnames(tbl))
 
   assert_that(length(mis) >= 1L)
 
-  res <- tbl[, c(ids, intersect(sta, colnames(tbl)), end)]
+  res <- load_src(tbl, cols = c(ids, intersect(sta, colnames(tbl)), end))
 
   if (length(mis) > 0L) {
     res[, c(mis) := 0L]
@@ -359,22 +364,32 @@ id_map_helper.default <- function(x, ...) stop_generic(x, .Generic)
 #'
 #' @seealso change_id
 #'
-#' @param x Passed to [as_id_cfg()] and [as_src_env()]
+#' @param x Data source (is coerced to `src_env` using `as_src_env()`)
+#'
+#' @return An `id_tbl` containing the selected IDs and depending on values
+#' passed as `in_time` and `out_time`, start and end times of the ID passed as
+#' `win_var`.
+#'
+#' @rdname stay_windows
+#' @export
+stay_windows <- function(x, ...) UseMethod("stay_windows", x)
+
 #' @param id_type Type of ID all returned times are relative to
 #' @param win_type Type of ID for which the in/out times is returned
 #' @param in_time,out_time column names of the returned in/out times
 #' @param interval The time interval used to discretize time stamps with,
 #' specified as [base::difftime()] object
 #' @param patient_ids Patient IDs used to subset the result
+#' @param ... Generic consistency
 #'
-#' @return An `id_tbl` containing the selected IDs and depending on values
-#' passed as `in_time` and `out_time`, start and end times of the ID passed as
-#' `win_var`.
-#'
+#' @rdname stay_windows
 #' @export
-stay_windows <- function(x, id_type = "icustay", win_type = id_type,
-                         in_time = "start", out_time = "end",
-                         interval = hours(1L), patient_ids = NULL) {
+stay_windows.src_env <- function(x, id_type = "icustay", win_type = id_type,
+                                 in_time = "start", out_time = "end",
+                                 interval = hours(1L), patient_ids = NULL,
+                                 ...) {
+
+  warn_dots(...)
 
   assert_that(is_interval(interval))
 
@@ -390,6 +405,41 @@ stay_windows <- function(x, id_type = "icustay", win_type = id_type,
 
   merge_patid(res, patient_ids)
 }
+
+#' @rdname stay_windows
+#' @export
+stay_windows.character <- function(x, ...) stay_windows(as.list(x), ...)
+
+#' @rdname stay_windows
+#' @export
+stay_windows.list <- function(x, ..., patient_ids = NULL) {
+
+  load_one <- function(x, ...) {
+
+    src <- src_name(x)
+    res <- stay_windows(x, ..., patient_ids = patient_ids[[src]])
+
+    if (mulit_src) {
+      res <- add_src_col(res, src)
+    }
+
+    res
+
+  }
+
+  x <- lapply(x, as_src_env)
+
+  srcs <- unique(chr_ply(x, src_name))
+  mulit_src <- length(srcs) > 1L
+
+  patient_ids <- split_patid(patient_ids, srcs)
+
+  rbind_lst(lapply(x, load_one, ...))
+}
+
+#' @rdname stay_windows
+#' @export
+stay_windows.default <- function(x, ...) stay_windows(as_src_env(x), ...)
 
 #' Switch between id types
 #'