+++ R/setup-download.R
@@ -33,13 +33,13 @@
 #'
 #' Demo datasets
 #' [MIMIC-III demo](https://physionet.org/content/mimiciii-demo/1.4/) and
-#' [eiCU demo](https://physionet.org/content/eicu-crd-demo/2.0/) can either be
+#' [eICU demo](https://physionet.org/content/eicu-crd-demo/2.0/) can either be
 #' installed as R packages directly by running
 #'
 #' ```
 #' install.packages(
 #'   c("mimic.demo", "eicu.demo"),
-#'   repos = "https://septic-tank.github.io/physionet-demo"
+#'   repos = "https://eth-mds.github.io/physionet-demo"
 #' )
 #' ```
 #'
@@ -355,16 +355,22 @@ download_pysionet_file <- function(url, dest = NULL, user = NULL,
     msg_ricu("Skipped download of {basename(url)}")
     return(invisible(NULL))
 
-  } else if (status == 401) {
+  } else if (status %in% c(401, 403)) {
 
     stop_ricu("Access to the requested resource was denied. Please set up an
                account at https://physionet.org/ and apply for data access.",
-              class = "physionet_401")
+              class = "physionet_login")
 
   } else if (status != 200) {
 
-    stop_ricu(rawToChar(res[["content"]]),
-              class = paste0("physionet_", status))
+    if (requireNamespace("xml2", quietly = TRUE)) {
+      info <- xml2::read_html(rawToChar(res[["content"]]))
+      info <- xml2::xml_text(xml2::xml_find_first(info, "//main"))
+    } else {
+      info <- "Unable to access requested resource on at {url}"
+    }
+
+    stop_ricu(info, class = paste0("physionet_", status))
   }
 
   if (head_only) {
@@ -455,7 +461,7 @@ download_check_data <- function(dest_folder, files, url, src,
 
   chksums <- tryCatch(
     get_sha256(url, user, pass),
-    physionet_401 = function(err) NULL
+    physionet_login = function(err) NULL
   )
 
   if (is.null(chksums)) {