+++ tests/testthat/test-concept.R
@@ -1,4 +1,49 @@
 
+test_that("load hirid items", {
+
+  gluc <- mockthat::with_mock(
+    load_id = id_tbl(id = c(20005110L, 24000523L, 24000585L),
+                     unit = rep("mmol/l", 3L)),
+    as_src_tbl = structure(list(),
+      class = c("hirid_tbl", "src_tbl"),
+      col_cfg = new_col_cfg("hirid", "observations", index_var = "datetime",
+                            time_vars = c("datetime", "entertime"),
+                            val_var = "value"),
+      src_env = structure(list(),
+        class = c("hirid_env", "src_env"),
+        id_cfg = new_id_cfg("hirid",  "icustay", "patientid", 1L,
+                            "admissiontime", NA_character_, "general")
+      )
+    ),
+    parse_dictionary(read_dictionary("concept-dict"), "hirid", "glu")
+  )
+
+  expect_identical(n_tick(gluc), 2L)
+
+  gluc <- as_item(gluc)[[1L]]
+
+  dat <- mockthat::with_mock(
+    load_ts = ts_tbl(
+      patientid = rep(seq_len(2L), each = 10L),
+      datetime = hours(rep(seq_len(10L), 2L)),
+      variableid = rep(24000585L, 20L),
+      value = rnorm(20L)
+    ),
+    do_itm_load(gluc)
+  )
+
+  expect_s3_class(dat, "ts_tbl")
+  expect_true(is_ts_tbl(dat))
+  expect_identical(data_vars(dat), c("variableid", "value", "unit"))
+  expect_equal(interval(dat), hours(1L))
+
+  expect_identical(id_vars(gluc), id_vars(dat))
+  expect_identical(index_var(gluc), index_var(dat))
+  expect_identical(meta_vars(gluc), meta_vars(dat))
+
+  expect_identical(dat, do_callback(gluc, dat))
+})
+
 skip_if_srcs_missing("mimic_demo")
 
 test_that("load concepts", {
@@ -99,6 +144,36 @@ test_that("load concepts", {
 
   expect_snapshot(print(gcs_con))
   expect_snapshot(print(gcs_raw))
+})
+
+skip_if_srcs_missing("eicu_demo")
+
+test_that("load concepts multi src", {
+
+  ids <- list(eicu_demo  = c(141765, 143870),
+              mimic_demo = c(293280, 298685))
+
+  dat <- load_concepts("glu", c("mimic_demo", "eicu_demo"),
+                       patient_ids = ids, verbose = FALSE)
+
+  expect_s3_class(dat, "ts_tbl")
+  expect_true(is_ts_tbl(dat))
+  expect_length(id_vars(dat), 2L)
+
+  uqe <- unique(dat[, id_vars(dat), with = FALSE])
+
+  expect_identical(nrow(uqe), 4L)
+  expect_identical(dat,
+    load_concepts("glu", c("mimic_demo", "eicu_demo"), patient_ids = uqe,
+                  verbose = FALSE)
+  )
+
+  uqe <- as.data.frame(uqe)
+
+  expect_identical(dat,
+    load_concepts("glu", c("mimic_demo", "eicu_demo"), patient_ids = uqe,
+                  verbose = FALSE)
+  )
 })
 
 test_that("load external dictionary", {