+++ tests/testthat/test-config.R
@@ -1,4 +1,6 @@
 
+skip_if_srcs_missing(c("mimic_demo", "eicu_demo"))
+
 test_that("config classes", {
 
   cfg <- load_src_cfg(c("mimic_demo", "eicu_demo"))